refactor: improve claim
